Bumped fastutils version to 8.5.2, JUnit to 4.13.1. Added new PZT versions for dealing with zeroes in the rating matrix.

+++ RankSys-novdiv/src/main/java/org/ranksys/novdiv/distance/VectorFeatureItemDistanceModel.java
@@ -51,7 +51,7 @@ public abstract class VectorFeatureItemDistanceModel<I, F> extends FeatureItemDi
         DoubleAdder norm1 = new DoubleAdder();
 
         features1.forEach(fv -> {
-            auxMap.put(fv.v1, fv.v2);
+            auxMap.put(fv.v1, (double) fv.v2);
             norm1.add(fv.v2 * fv.v2);
         });
